+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/container/workspace-interactions/workspace-interactions.tsx
@@ -16,7 +16,7 @@ import * as React from 'react'
 import WorkspaceInteractionsPresentation from '../../presentation/workspace-interactions'
 import { hot } from 'react-hot-loader'
 import withListenTo, { WithBackboneProps } from '../backbone-container'
-import { Sharing } from '../sharing'
+import { Sharing } from '../../sharing'
 import { Security, Restrictions } from '../../utils/security'
 const user = require('../../../component/singletons/user-instance.js')
 const store = require('../../../js/store.js')
