+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/sharing/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { Sharing, Item, Category } from './sharing'
+export { Sharing, Item, Category } from './container'
