+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/sharing/presentation.tsx
@@ -15,11 +15,11 @@
 import * as React from 'react'
 
 import { hot } from 'react-hot-loader'
-import Text from '../../container/input-wrappers/text/text'
-import Enum from '../../container/input-wrappers/enum/enum'
-import { Category, Item } from '../../container/sharing'
+import Text from '../container/input-wrappers/text/text'
+import Enum from '../container/input-wrappers/enum/enum'
+import { Category, Item } from '.'
 import styled from 'styled-components'
-import { Access } from '../../utils/security'
+import { Access } from '../utils/security'
 
 const Root = styled.div`
   position: relative;
